Lexer in C almost ready (should be exciting to see if there are any speed improvements)
