+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "C_Cpp.errorSquiggles": "Disabled"
+}
