+++ build.bat
@@ -2,6 +2,6 @@ mkdir bin
 
 pushd bin
 
-g++ ../src/ppl.cpp
+g++ ../src/ppl.cpp -I ../src/
 
 popd
